Allow "--verbose" without to explicit boolean flag to enable verbose changelog listings (#208)
Yargs isn't set to interpret the --verbose flag as a boolean, meaning  "--verbose true" is required to enable verbose listings instead of just "--verbose". The latter will be accepted, but will somewhat confusingly not enable verbose listings. Tell Yargs the type of the verbose flag is a boolean to support both forms.