Move the releases repo to a monorepo
This moves the current repo to a yarn workspaces based monorepo with separate packages for the changelog generator any cherry-pick script. You can still run either without build by running:

`yarn generate` for changelog generator or
`yarn cherry-pick` (for cherry-pick script)

I pulled in some of the RNW repo infra packages, which enable ESLint with static analysis rules on any TypeScript project. I also brought in "lage", which is a similar task runner to "lerna", but will cache results to avoid things like unneccesary lint or rebuild on changes.

THis also brings in `just-scripts` which allows creation of build-or-lint tasks. E.g. we use it to define linting the changelog creation projectr as running eslint, then running tsc to typecheck.